+++ apps/docs-website/docusaurus.config.ts
@@ -21,6 +21,10 @@ const config: Config = {
   // For GitHub pages deployment, it is often '/<projectName>/'
   baseUrl: '/',
 
+  future: {
+    experimental_faster: true,
+  },
+
   // GitHub pages deployment config.
   // If you aren't using GitHub pages, you don't need these.
   organizationName: 'AndreaPontrandolfo', // Usually your GitHub org/user name.
@@ -49,6 +53,9 @@ const config: Config = {
         },
         blog: {
           showReadingTime: true,
+          feedOptions: {
+            xslt: true,
+          },
           remarkPlugins: [
             [require('@docusaurus/remark-plugin-npm2yarn'), { sync: true }],
           ],