+++ packages/eslint-config-sheriff/src/playwrightConfig.ts
@@ -2,15 +2,15 @@ import playwright from 'eslint-plugin-playwright';
 import { allJsExtensions } from '@sherifforg/constants';
 import { playwrightHandPickedRules } from './handpickedRules/playwrightHandPickedRules';
 
-export const playwrightConfig = [
-  {
-    ...playwright.configs['flat/recommended'],
-    files: [`**/*{${allJsExtensions}}`],
-  },
-  {
-    files: [`**/*{${allJsExtensions}}`],
-    rules: {
-      ...playwrightHandPickedRules,
+export const getPlaywrightConfig = (pathsOverrides?: string[]) => {
+  return [
+    {
+      ...playwright.configs['flat/recommended'],
+      files: pathsOverrides ?? [`**/*{${allJsExtensions}}`],
     },
-  },
-];
+    {
+      files: pathsOverrides ?? [`**/*{${allJsExtensions}}`],
+      rules: playwrightHandPickedRules,
+    },
+  ];
+};