+++ packages/sheriff-create-config/src/utils/throwError.ts
@@ -8,7 +8,7 @@ interface Error {
 export const throwError = (message: string, { error }: Error = {}): void => {
   if (error) {
     consola.error(message);
-    throw new Error(isString(error) ? error : String(error));
+    throw new Error(isString(error) ? error : JSON.stringify(error));
   }
 
   throw new Error(message);