+++ packages/eslint-config-sheriff/CHANGELOG.md
@@ -1,5 +1,23 @@
 # eslint-config-sheriff
 
+## 25.1.0
+
+### Minor Changes
+
+- ce376df: feat(rules): updated eslint-plugin-import and allowed some rules in `*.config.*` files
+  Fixes #281
+- 527bedd: feature(rules): allow side-effects imports in no-restricted-imports
+
+### Patch Changes
+
+- 09f4162: feature(deps): updated eslint-plugin-react-hooks to V5
+
+## 25.0.0
+
+### Major Changes
+
+- 44643a6: feat: introduced dts-bundle-generator to bundle @sherifforg/types
+
 ## 24.0.0
 
 ### Major Changes